+++ functions.php
@@ -23,7 +23,7 @@ final class BFG {
 
 	public static function get_instance() {
 
-		if( null === self::$instance )
+		if( self::$instance === null )
 			self::$instance = new self();
 
 		return self::$instance;