+++ includes/developer-tools.php
@@ -15,19 +15,19 @@ function bfg_clear_transients_node($wp_admin_bar) {
 
 	global $wpdb;
 
-	if( isset($_GET['clear-transients']) && 1 === (int) $_GET['clear-transients'] ) {
+	if( isset($_GET['clear-transients']) && (int) $_GET['clear-transients'] === 1 ) {
 		$wpdb->query( "DELETE FROM `{$wpdb->options}` WHERE `option_name` LIKE ('_transient_%') OR `option_name` LIKE ('_site_transient_%')" );
 		wp_cache_flush();
 		add_action( 'admin_notices', 'bfg_transients_cleared_notice' );
 	}
 
-	if( isset($_GET['clear-orphans']) && 1 === (int) $_GET['clear-orphans'] ) {
+	if( isset($_GET['clear-orphans']) && (int) $_GET['clear-orphans'] === 1 ) {
 		$wpdb->query( "DELETE pm FROM `{$wpdb->postmeta}` pm LEFT JOIN `{$wpdb->posts}` wp ON wp.ID = pm.post_id WHERE wp.ID IS NULL;" );
 		$wpdb->query( "DELETE um FROM `{$wpdb->usermeta}` um LEFT JOIN `{$wpdb->users}` wp ON wp.ID = um.user_id WHERE wp.ID IS NULL;" );
 		add_action( 'admin_notices', 'bfg_orphans_cleared_notice' );
 	}
 
-	if( isset($_GET['clear-ipq']) && 1 === (int) $_GET['clear-ipq'] ) {
+	if( isset($_GET['clear-ipq']) && (int) $_GET['clear-ipq'] === 1 ) {
 		$wpdb->query("TRUNCATE TABLE {$wpdb->prefix}queue_failures;");
 		$wpdb->query("TRUNCATE TABLE {$wpdb->prefix}queue_jobs;");