+++ includes/structure/search.php
@@ -43,7 +43,7 @@ function bfg_redirect_single_search_result() {
 	if( is_search() ) {
 		global $wp_query;
 
-		if( 1 === $wp_query->post_count) {
+		if( $wp_query->post_count === 1) {
 			wp_safe_redirect( get_permalink( $wp_query->posts['0']->ID ) );
 			exit;
 		}